Remove unneeded ImageController
